Introduce ResourceDecoratorBase
