+++ src/Plugin/resource/Decorators/CachedResource.php
@@ -2,25 +2,19 @@
 
 /**
  * @file
- * Contains \Drupal\restful\Plugin\resource\CachedResource
+ * Contains \Drupal\restful\Plugin\resource\Decorators\CachedResource
  */
 
-namespace Drupal\restful\Plugin\resource;
+namespace Drupal\restful\Plugin\resource\Decorators;
 
-use Drupal\Component\Plugin\PluginBase;
 use Drupal\restful\Http\HttpHeader;
-use Drupal\restful\Http\RequestInterface;
+use Drupal\restful\Plugin\resource\DataProvider\CachedDataProvider;
+use Drupal\restful\Plugin\resource\DataProvider\CachedDataProviderInterface;
 use Drupal\restful\Plugin\resource\DataProvider\DataProviderInterface;
+use Drupal\restful\Plugin\resource\ResourceInterface;
 use Drupal\restful\Resource\ResourceManager;
 
-class CachedResource extends PluginBase implements ResourceInterface {
-
-  /**
-   * The decorated resource.
-   *
-   * @var ResourceInterface
-   */
-  protected $subject;
+class CachedResource extends ResourceDecoratorBase implements CachedResourceInterface {
 
   /**
    * Cache controller object.
@@ -53,10 +47,7 @@ class CachedResource extends PluginBase implements ResourceInterface {
   }
 
   /**
-   * Getter for $cacheController.
-   *
-   * @return \DrupalCacheInterface
-   *   The cache object.
+   * {@inheritdoc}
    */
   public function getCacheController() {
     return $this->cacheController;
@@ -103,61 +94,17 @@ class CachedResource extends PluginBase implements ResourceInterface {
   /**
    * {@inheritdoc}
    */
-  public function dataProviderFactory() {
-    return $this->subject->dataProviderFactory();
-  }
-
-  /**
-   * Proxy method to get the account from the rateLimitManager.
-   *
-   * {@inheritdoc}
-   */
-  public function getAccount($cache = TRUE) {
-    return $this->subject->getAccount();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getRequest() {
-    return $this->subject->getRequest();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getPath() {
-    return $this->subject->getPath();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setPath($path) {
-    $this->subject->setPath($path);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getFieldDefinitions() {
-    return $this->subject->getFieldDefinitions();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   public function getDataProvider() {
     // Make sure that this returns a cached data provider.
     if (isset($this->dataProvider)) {
       return $this->dataProvider;
     }
     $data_provider = $this->dataProviderFactory();
-    if ($data_provider instanceof DataProvider\CachedDataProvider) {
+    if ($data_provider instanceof CachedDataProviderInterface) {
       $this->dataProvider = $data_provider;
       return $this->dataProvider;
     }
-    $this->dataProvider = new DataProvider\CachedDataProvider($data_provider, $this->getCacheController());
+    $this->dataProvider = new CachedDataProvider($data_provider, $this->getCacheController());
     $plugin_definition = $this->subject->getPluginDefinition();
     $this->dataProvider->addOptions(array(
       'renderCache' => $this->defaultCacheInfo(),
@@ -175,38 +122,10 @@ class CachedResource extends PluginBase implements ResourceInterface {
   /**
    * {@inheritdoc}
    */
-  public function getResourceName() {
-    return $this->subject->getResourceName();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   public function process() {
     $path = $this->getPath();
 
     return ResourceManager::executeCallback($this->getControllerFromPath($path), array($path));
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function controllersInfo() {
-    return $this->subject->controllersInfo();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getControllers() {
-    return $this->subject->getControllers();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function index($path) {
-    return $this->getDataProvider()->index();
   }
 
   /**
@@ -224,97 +143,6 @@ class CachedResource extends PluginBase implements ResourceInterface {
       ->add(HttpHeader::create('Link', $this->versionedUrl($canonical_path, array(), FALSE) . '; rel="canonical"'));
 
     return $this->getDataProvider()->viewMultiple($ids);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function create($path) {
-    return $this->subject->create($path);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function update($path) {
-    return $this->subject->update($path);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function replace($path) {
-    return $this->subject->replace($path);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function remove($path) {
-    $this->subject->remove($path);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getVersion() {
-    return $this->subject->getVersion();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function versionedUrl($path = '', $options = array(), $version_string = TRUE) {
-    return $this->subject->versionedUrl($path, $options, $version_string);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getConfiguration() {
-    return $this->subject->getConfiguration();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setConfiguration(array $configuration) {
-    $this->subject->setConfiguration($configuration);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function defaultConfiguration() {
-    return $this->subject->defaultConfiguration();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function calculateDependencies() {
-    return $this->subject->calculateDependencies();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setRequest(RequestInterface $request) {
-    $this->subject->setRequest($request);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function access() {
-    return $this->subject->access();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getControllerFromPath($path = NULL, ResourceInterface $resource = NULL) {
-    return $this->subject->getControllerFromPath($path, $resource ?: $this);
   }
 
   /**
